+++ v2/configuration.go
@@ -10,6 +10,14 @@ import (
 	"strings"
 )
 
+const (
+	HUB_PREFIX       = "00000"
+	HUB_NOTIFY       = "https://notify.insighthub.smartbear.com"
+	HUB_SESSION      = "https://sessions.insighthub.smartbear.com"
+	DEFAULT_NOTIFY   = "https://notify.bugsnag.com"
+	DEFAULT_SESSIONS = "https://sessions.bugsnag.com"
+)
+
 // Endpoints hold the HTTP endpoints of the notifier.
 type Endpoints struct {
 	Sessions string
@@ -209,18 +217,35 @@ func (config *Configuration) IsAutoCaptureSessions() bool {
 }
 
 func (config *Configuration) updateEndpoints(endpoints *Endpoints) {
+	sessionsDisabled := false
 	if endpoints.Notify != "" {
 		config.Endpoints.Notify = endpoints.Notify
 		if endpoints.Sessions == "" {
 			config.Logger.Printf("WARNING: Bugsnag notify endpoint configured without also configuring the sessions endpoint. No sessions will be recorded")
 			config.Endpoints.Sessions = ""
+			sessionsDisabled = true
+		}
+	} else {
+		if strings.HasPrefix(config.APIKey, HUB_PREFIX) {
+			config.Endpoints.Notify = HUB_NOTIFY
+		} else {
+			config.Endpoints.Notify = DEFAULT_NOTIFY
 		}
 	}
+
 	if endpoints.Sessions != "" {
 		if endpoints.Notify == "" {
 			panic("FATAL: Bugsnag sessions endpoint configured without also changing the notify endpoint. Bugsnag cannot identify where to report errors")
 		}
 		config.Endpoints.Sessions = endpoints.Sessions
+	} else {
+		if !sessionsDisabled {
+			if strings.HasPrefix(config.APIKey, HUB_PREFIX) {
+				config.Endpoints.Sessions = HUB_SESSION
+			} else {
+				config.Endpoints.Sessions = DEFAULT_SESSIONS
+			}
+		}
 	}
 }