+++ v2/event.go
@@ -142,7 +142,9 @@ func newEvent(rawData []interface{}, notifier *Notifier) (*Event, *Configuration
 			event.Stacktrace = make([]StackFrame, len(err.StackFrames()))
 
 		case bool:
-			config = config.merge(&Configuration{Synchronous: bool(datum)})
+			// clone without merging to avoid double-call to Configuration.update
+			config = config.clone()
+			config.Synchronous = bool(datum)
 
 		case severity:
 			event.Severity = datum