+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## TBD
+
+### Enhancements
+
+* Set default endpoints based on API key [#259](https://github.com/bugsnag/bugsnag-go/pull/259)
+
 ## 2.5.1 (2024-10-21)
 
 ### Bug fixes